requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
openai>=1.3.0
python-dotenv>=1.0.0
pyaudio>=0.2.13
pydub>=0.25.1
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0 